add an extra implicit param to User#find_pseudonym_for_account and Pseudonym#works_for_account?
refs #6883

Use it when checking if a pseudonym is valid for authentication, but
not for if we need to create a new pseudonym for an account.  This
means that you can authenticate via an implicit trust relationship,
but when you are invited to something, it will still try and create
a pseudonym for you in that account.

test plan:
 * run existing specs (no behavior is changed without a plugin
   to define trust relationships)

Change-Id: Ieb200da27351c185e640e9c50ae1a6054e502c63
Reviewed-on: https://gerrit.instructure.com/8076
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>