don't show the filter until everything in the scope is filterable
since the data migration will take a long time to run, we don't want to
show the filter dropdown thingy until it can be used. along with that,
we now differentiate between not-yet-tagged (null) and no tags (empty
string).

also tweaked it so we show the computed default tags if the conversation
is not-yet-tagged. this way the transition will be seamless from a ui
point of view (there will never be a period where we don't show common
contexts in the ui)

test plan:
 1. clear out all tags on conversations, conversation_participants, and
    conversation_message_participants
 2. confirm that common contexts still appear in the ui, but the filter
    is not visible
 3. re-run the tag migration
 4. confirm that the filter appears once the migration completes (it
    could appear slightly before it's done ... basically as soon as all
    *your* conversations are tagged)

Change-Id: I6df6b0653bea9d306560917c29ff31625a7b6d93
Reviewed-on: https://gerrit.instructure.com/8340
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>