observers see all link enrollment grades for course; closes #6441
if an observer is linked to two students in the same course, we want to provide
an option on the grades page to choose which students grades to view, and be
able to switch back and forth between them.  We also want to handle these
multiple links on the course settings users tab.

test plan:
- create a course with > 2 students and 1 observer
- creates > 2 enrollments for the observer, each linked to a different student
- go to the course settings users tab, make sure the multiple links show up and
  you can change/remove/add links as you would expect.
- go to the course grades page. You'll see a dropdown allowing you to choose
  which student's grades to view.

Change-Id: Ia79cb99c0bfd54924d0aee1f8549a4e82121abae
Reviewed-on: https://gerrit.instructure.com/8481
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>