fix assignment status in 'Coming Up' sidebar. fixes #6783
submitted assignments weren't having their status consistently
updated because the partial depended on the @current_user_submissions
variable that wasn't being set in every view.

fixed this by adding an extra conditional to the view (eek!). also
included a fix to ensure that the correct context is always used
when checking permissions (the page context isn't always the context
of the 'Coming Up' item).

test plan:
  * create a course with an assignment due tomorrow;
  * as student of that course, log in and submit the assignment;
  * verify that the assignment shows as submitted on both the user
    dashboard and the course home page.
  * click around the app and ensure that the sidebar works as
    expected.

Change-Id: If1b955cd4e49c3c83cae9c92a0f3c515e8d3c2d3
Reviewed-on: https://gerrit.instructure.com/8633
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Ryan Florence <ryanf@instructure.com>