link student names to their grades page in gradebook2
fixes #6933
refs #6809

I added html_url and grades.html_url to to enrollments response so we
can pull this onto the page. As part of this, I changed which API
gradebook2 uses for pulling the student enrollment data -- it now uses
the enrollments API, which is the way going forward. Including students
in the course sections API will probably get deprecated at some point.

As a consequence, we are now part of the way to being able to choose
whether to filter concluded enrollments.

test plan:
  * go to gradebook2, verify that clicking a student's name takes you to
    the student's individual grades page

Change-Id: I7c043fda08500a684bfeb8391da012b653a0f6cf
Reviewed-on: https://gerrit.instructure.com/8927
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Hudson <hudson@instructure.com>