Allow launch preference when external tools have matching precedence
It's possible to have two external tools with all the
same configuration except for custom fields. In this case,
inserting two links, one to each tool configuration, would
have resulted in two launches of the same tool because
Canvas was only searching for the first match. This
commit adds an additional "preference" for the specified
tool, so Canvas doesn't just pick the first one, it tries
to pick the preferred one if there's nothing else
trumping that tool in precedence.

This does not fix the problem with having a tool's custom fields
overridden by a different tool that has higher precedence.

fixes #6797

test plan:
- add two external tools to a course with the same configuration
  but different custom fields
- add a link to each tool in the course modules
- make sure the two links send over the correct custom fields,
  those matching the tool that was picked when inserting the link

Change-Id: I8b037f169cfcfbfd9f9082a9dbec6fcee442edaa
Reviewed-on: https://gerrit.instructure.com/9025
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>