export CC html content as utf-8
We didn't include any Content-Type meta tag, so on import Nokogiri was
just guessing at the encoding (and getting it wrong).

Another thing we could do is default to utf-8 on CC import, if no
content-type is specified -- but that's pretty scary, I wouldn't be
surprised if there's a lot of ISO-8859-1 CC packages out there.

fixes #7066

test plan:
  * export a course that has rich content such as a wiki page with
    extended utf-8 characters, such as a literal nbsp ("\302\240"), or
    any other characters not in ISO-8859-1
  * re-import that package, and verify that your text was not munged

Change-Id: I14d01e9165c05cfada2bcc3625b0ae6e29cef022
Reviewed-on: https://gerrit.instructure.com/9045
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>