return attach permission on discussion topics, closes #7376
This way api users can know whether the current user can post
attachments to the discussion or not.

test plan: hit the discussions api as a teacher in the course, verify
that attach is always true. hit it as a student, verify attach is only
true if the course settings allow student attachments.

Change-Id: I508e6b8f329f8472e0a2899676871a1b5318f585
Reviewed-on: https://gerrit.instructure.com/9384
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Stephen Lottermoser <stephen@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>