add new 'external' avatar image type; closes #7763
support for a new type of avatar image, which is simply a url
pointing to an externally hosted image. the use case is that schools
who host there own repository of user images want to simply refer to
these image for canvas profile pictures.

test-plan:
this features is currently only available via the api, so try
setting your profile picture to an external URL via the api, and make
sure it shows up correctly. also try setting your avatar to something
that is larger than 50px and non-square (we're encouraging people to
use 50px square images, but not enforcing it).

Change-Id: I5a1de6159d6e5fd490dd59871e98d823d0177d7a
Reviewed-on: https://gerrit.instructure.com/9665
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>