change ajax url for calendar1 month data
Originally the calendar was at e.g. /courses/X/calendar so monthDataURL
contained the specific calendar endpoint to use, but now the calendar is
always at /calendar, so it's no longer necessary. We're removing it
because it was using params in an unsafe manner.

refs #8075

test plan: this should have no visible change in behavior -- verify that
the calendar still works and you can still switch between months

Change-Id: I2d2ac63f0beb66cee0092f7b5c8fe79d32f1c573
Reviewed-on: https://gerrit.instructure.com/9894
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>