don't use eval in error handler
fixes #8219

typically $.parseJSON will be available here, and if it's not, we just
won't be able to parse a JSON formatted error response. We'll still have
the text of the error response available though.

test plan: in normal usage this will have no visible effect. if you
crafted a malicious XHR response with executable javascript, you could
verify that it's no longer executed in the error handler.

Change-Id: Ibc42a7d7ce1edc930a884d2a79013a1c9e052080
Reviewed-on: https://gerrit.instructure.com/10122
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>