grant read permission to teachers of date-concluded, unpublished course
fixes #7484, and also allows teachers to view the content of these
date-concluded, unpublished courses, which they couldn't before.

test plan: as a teacher who is *not* also an admin, visit a course that
was never published, but is now "soft" concluded because the term it's in is
concluded. verify that you can view course content, and copy the course
content into another course.

Change-Id: I8af6213e8f2ecf0ffc758acee0080e6ffbf9b9dc
Reviewed-on: https://gerrit.instructure.com/10163
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>