create media object if it doesn't exist, fixes #8420
when adding a media comment with a conversation message, don't require the
media object to exist beforehand. this is important since there is no
(published) media object endpoint, so mobile apps (and other api users)
need to set the media comment info after uploading to kaltura. this brings
conversations in line with how we do things elsewhere.

test plan:
1. create a conversation message against the api, setting a media comment
   that doesn't already have a media_object
2. it should work, and the media_object should get created

Change-Id: Ibf1b1e417e85288c19614af66a9960b738dc8813
Reviewed-on: https://gerrit.instructure.com/10488
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Cameron Matheson <cameron@instructure.com>