properly escape discussion reply text with attachment; fixes #8509
when you submit an attachment with a discussion reply, we use a custom backbone
sync method in order to support multipart.  to do this we build a new form in
an iframe and insert it into the page, and values weren't getting escaped
properly when they were inserted into the new form.

test plan:
- in a new discussion, entry a reply that includes an apostrophe ', and angle
  brackets < and >. also attach a file to the reply.
- your message should come through in its entirety.

Change-Id: Iecb262f38819c184be4187891d7a30d17a8d80dd
Reviewed-on: https://gerrit.instructure.com/11050
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>